requirements.txt
OpenAI
pandas
python-dotenv
langchain
langchain_community
langchain_openai
black
chromadb
pypdf